[geolocation] Leave a chance to all location providers to get a fix
In our case, the NetworkLocationProvider will always fail
for lack of a valid API token, and we don't want that to
take precedence over the QtPositioning-based backend.

And disable location requests to the geolocation google API.
If we don't have a valid token, it will fail anyway.
And for now we don't provide a way to use an existing API token.

Change-Id: Ic175bd3fb527a76a578ef3568f7ac7ed07c4ccad
Reviewed-by: Michael Brüning <michael.bruning@qt.io>
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>