[gn-config] Add toolchain script for msvc setup
This script is a simplified and adjusted version
of setup_toolchains.py. Uses the same idea to
create environment files accroding to
https://learn.microsoft.com/en-us/windows/win32/procthread/environment-variables

The script runs vcvarsall.bat and captures the environment
setup, so it can be used for msvc toolchain setup.

The initial approch to have this functionaly on cmake side
failed as cmake is not able to handle nul characters
which are used as terminators (see the link above).

Task-number: QTBUG-129970
Change-Id: I02a4708b4718f1813932f4faa6ff7091f1292c62
Reviewed-on: https://codereview.qt-project.org/c/qt/qtwebengine-chromium/+/618113
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>