Fix recursion/deadlock in partition_alloc
The internal allocator used for PartitionAllocator's own allocations
is supposed to avoid reentrancy issues, but recent changes made it
recurse into itself when initializing. More specifically, the
PartitionRoot class has a member, scheduler_loop_quarantine,
which contains an std::vector that's using the internal allocator.
So when initializing, it ends up recursing into
partition_alloc::internal::InternalAllocatorRoot(), where we have
a static PartitionRoot instance.

This all works fine on most platforms, but MSVC debug builds have extra
checks for static variables that stall the thread until the static
is fully constructed. But since we're recursing into the function,
the process just ends up waiting forever.

To remove the recursion, this change removes the use of
PartitionAllocator from the offending std::vector, and makes it
fall back to the default one.

Change-Id: I123ba980d6e776905d65f5b97e904e1f2e3b8c61
Reviewed-on: https://codereview.qt-project.org/c/qt/qtwebengine-chromium/+/697864
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>
Reviewed-by: Peter Varga <pvarga@inf.u-szeged.hu>